Move BukkitChunkData out of BukkitChunkGenerator
